Crate solana_address

Address representation for Solana.

An address is a sequence of 32 bytes, often shown as a base58 encoded string (e.g. 14grJpemFaf88c8tiVb77W7TYg2W3ir6pfkKz3YjhhZ5).

Modules

error

syscalls

Macros

address

Convenience macro to define a static Address value.

declare_deprecated_id

Same as declare_id except that it reports that this ID has been deprecated.

declare_id

Convenience macro to declare a static address and functions to interact with it.

Structs

Address

The address of a Solana account.

AddressHasher

A faster, but less collision resistant hasher for addresses.

AddressHasherBuilder

A builder for faster, but less collision resistant hasher for addresses.

Constants

ADDRESS_BYTES

Number of bytes in an address.

MAX_SEEDS

Maximum number of seeds

MAX_SEED_LEN

maximum length of derived Address seed

PDA_MARKER

Marker used to find program derived addresses (PDAs).

Functions

address_eq

Custom implementation of equality for Address.

bytes_are_curve_point